CXP-1165: move User folder to Service
