+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/Apps/handlers.yml
@@ -4,7 +4,7 @@ services:
             - '@validator'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Session\AppAuthorizationSession'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Marketplace\Persistence\GetAppQuery'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\User\Internal\CreateUser'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Service\User\CreateUser'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\User\CreateUserGroup'
             - '@Akeneo\Connectivity\Connection\Application\Apps\Service\CreateConnection'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\AppRoleWithScopesFactory'
@@ -43,10 +43,10 @@ services:
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\GetAppDeletionQuery'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\DeleteConnectedAppQuery'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\User\Internal\DeleteUser'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Service\User\DeleteUser'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Client\Fos\DeleteClient'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\User\Internal\DeleteUserGroup'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\User\Internal\DeleteUserRole'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Service\User\DeleteUserGroup'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Service\User\DeleteUserRole'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\SaveRevokedAccessTokensOfDisconnectedAppQuery'
 
     Akeneo\Connectivity\Connection\Application\Apps\Command\RefreshConnectedAppHandler:
