+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Service/User/DeleteUserGroupIntegration.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\back\tests\Integration\User\Internal;
+namespace Akeneo\Connectivity\Connection\Tests\Integration\Service\User;
 
-use Akeneo\Connectivity\Connection\Infrastructure\User\Internal\DeleteUserGroup;
+use Akeneo\Connectivity\Connection\Infrastructure\Service\User\DeleteUserGroup;
 use Akeneo\Connectivity\Connection\Tests\CatalogBuilder\Enrichment\UserGroupLoader;
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
